+++ index.html
@@ -231,7 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-<script src="main.js">
+<script src="main.js" type="module">
 </script>
 </body>
 </html>
